Moved identifier and tags updaters into sub-folder
* This matches where other SecondaryChannel updaters were placed
