+++ src/managers/channelManager/shared/SecondaryChannelEmail.ts
@@ -6,8 +6,8 @@ import OneSignalApi from "../../../OneSignalApi";
 import Database from "../../../services/Database";
 import { SecondaryChannel, SecondaryChannelWithControllerEvents } from "./SecondaryChannel";
 import { SecondaryChannelController } from "./SecondaryChannelController";
-import { SecondaryChannelIdentifierUpdater } from "./SecondaryChannelIdentifierUpdater";
-import { SecondaryChannelExternalTagsUpdater } from "./SecondaryChannelTagsUpdater";
+import { SecondaryChannelIdentifierUpdater } from "./updaters/SecondaryChannelIdentifierUpdater";
+import { SecondaryChannelExternalTagsUpdater } from "./updaters/SecondaryChannelTagsUpdater";
 import { SecondaryChannelExternalUserIdUpdater } from "./updaters/SecondaryChannelExternalUserIdUpdater";
 import { SecondaryChannelFocusUpdater } from "./updaters/SecondaryChannelFocusUpdater";
 import { SecondaryChannelSessionUpdater } from "./updaters/SecondaryChannelSessionUpdater";
